+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.34.0
 
-- @TODO
+- implement [@Authorize() decorator](https://egomobile.github.io/node-http-server/modules.html#Authorize) and [authorize](https://egomobile.github.io/node-http-server/interfaces/IControllerRouteOptions.html#authorize) options
 
 ## 0.33.1